requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.2
Automat==0.7.0
billiard==3.6.0.0
celery==4.3.0
certifi==2018.11.29
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
crypto==1.4.1
cryptography==2.7
daphne==2.3.0
Django==2.1
django-cors-headers==2.4.0
django-extensions==2.1.6
django-rest-swagger==2.2.0
django-stubs==0.12.1
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
djangorestframework-stubs==0.3.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.5.0
MarkupSafe==1.1.0
msgpack==0.6.1
mypy==0.701
mypy-extensions==0.4.1
Naked==0.1.31
openapi-codec==1.3.2
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.19
pycrypto==2.6.1
pyfcm==1.4.7
PyHamcrest==1.9.0
PyJWT==1.7.1
pypcapfile==0.12.0
python-magic==0.4.15
pytz==2018.9
PyYAML==5.1
redis==3.3.8
requests==2.21.0
shellescape==3.4.1
simplejson==3.16.0
six==1.12.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.3.5
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
zope.interface==4.6.0
